backend/requirements.txt
Django==3.2.3
djangorestframework==3.14.0
djoser==2.2.2
Pillow==9.2.0
django-extra-fields==3.0.2
psycopg2-binary==2.9.3
django-filter==23.5
PyYAML==6.0
gunicorn==20.1.0
python-dotenv==0.19.0